Move file changes to snapshotting subproject
