+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/ErrorHandlingTaskStateChanges.java
@@ -18,6 +18,7 @@ package org.gradle.api.internal.changedetection.rules;
 
 import org.gradle.api.GradleException;
 import org.gradle.api.Task;
+import org.gradle.internal.changes.TaskStateChangeVisitor;
 
 public class ErrorHandlingTaskStateChanges implements TaskStateChanges {
     private final Task task;
