+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/InputFileTaskStateChanges.java
@@ -19,6 +19,7 @@ package org.gradle.api.internal.changedetection.rules;
 import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.api.NonNullApi;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
+import org.gradle.internal.changes.TaskStateChangeVisitor;
 import org.gradle.internal.fingerprint.FileCollectionFingerprint;
 
 @NonNullApi
