+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/InputPropertyValueTaskStateChanges.java
@@ -22,6 +22,7 @@ import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.changedetection.state.ImplementationSnapshot;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
 import org.gradle.api.internal.changedetection.state.ValueSnapshot;
+import org.gradle.internal.changes.TaskStateChangeVisitor;
 
 import java.util.Map;
