+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/PreviousSuccessTaskStateChanges.java
@@ -18,6 +18,8 @@ package org.gradle.api.internal.changedetection.rules;
 
 import org.gradle.api.NonNullApi;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
+import org.gradle.internal.changes.TaskStateChange;
+import org.gradle.internal.changes.TaskStateChangeVisitor;
 
 @NonNullApi
 public class PreviousSuccessTaskStateChanges implements TaskStateChanges {
