+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/SummaryTaskStateChanges.java
@@ -16,6 +16,9 @@
 
 package org.gradle.api.internal.changedetection.rules;
 
+import org.gradle.internal.changes.TaskStateChange;
+import org.gradle.internal.changes.TaskStateChangeVisitor;
+
 import java.util.Arrays;
 import java.util.List;
