+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/TaskTypeTaskStateChanges.java
@@ -20,6 +20,7 @@ import com.google.common.collect.ImmutableList;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.changedetection.state.ImplementationSnapshot;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
+import org.gradle.internal.changes.TaskStateChangeVisitor;
 
 import javax.annotation.Nullable;
