+++ subprojects/core/src/test/groovy/org/gradle/api/internal/changedetection/rules/SummaryTaskStateChangesTest.groovy
@@ -17,6 +17,7 @@
 
 package org.gradle.api.internal.changedetection.rules
 
+import org.gradle.internal.changes.TaskStateChange
 import spock.lang.Specification
 
 class SummaryTaskStateChangesTest extends Specification {
