+++ subprojects/core/src/test/groovy/org/gradle/internal/fingerprint/impl/EmptyFileCollectionFingerprintTest.groovy
@@ -18,8 +18,8 @@ package org.gradle.internal.fingerprint.impl
 
 import com.google.common.collect.ImmutableMultimap
 import org.gradle.api.internal.changedetection.rules.CollectingTaskStateChangeVisitor
-import org.gradle.api.internal.changedetection.rules.FileChange
-import org.gradle.api.internal.changedetection.rules.TaskStateChange
+import org.gradle.internal.changes.FileChange
+import org.gradle.internal.changes.TaskStateChange
 import org.gradle.internal.file.FileType
 import org.gradle.internal.fingerprint.FileCollectionFingerprint
 import org.gradle.internal.hash.HashCode
