+++ subprojects/snapshots/src/main/java/org/gradle/internal/changes/ChangeType.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.changedetection.rules;
+package org.gradle.internal.changes;
 
 public enum ChangeType {
     ADDED("has been added"),
