+++ subprojects/snapshots/src/main/java/org/gradle/internal/changes/TaskStateChange.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.changedetection.rules;
+package org.gradle.internal.changes;
 
 public interface TaskStateChange {
     String getMessage();
